Refactor reticulumd inbound worker modules
